requirements.txt
# Core
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0

# SerpAPI
google-search-results>=2.4.2

# Web Scraping (Amazon, Exploding Topics)
beautifulsoup4>=4.12.0
lxml>=4.9.0
fake-useragent>=1.4.0

# UI
streamlit>=1.29.0

# Data Processing
thefuzz>=0.20.0  # Fuzzy string matching for merger
python-Levenshtein>=0.23.0  # Speed up thefuzz

# Existing (Google Trends - legacy)
pytrends>=4.9.2
